début d'un clean du index.js
